Code update
PiperOrigin-RevId: 945779496
